+++ book_store_be/src/main/java/com/project/book_store_be/Repository/AuthorRepository.java
@@ -8,6 +8,6 @@ import javax.swing.text.html.Option;
 import java.util.Optional;
 
 @Repository
-public interface AuthorRepository extends JpaRepository<Author,Integer> {
+public interface AuthorRepository extends JpaRepository<Author,Long> {
     Optional<Author> findByPseudonym(String pseudonym);
 }
